Added avatar screen with icons
Added avatar screen with icons
